// src/types/deprecated/tribes.ts
import { type CastawayName } from '~/types/deprecated/castaways';
import { type EpisodeNumber } from '~/types/deprecated/episodes';
import { type SeasonName } from '~/types/deprecated/seasons';

export type TribeId = number;
export type TribeName = string;
export type TribeColor = string;

export interface Tribe {
  tribeId: TribeId;
  tribeName: TribeName;
  tribeColor: TribeColor;
  seasonName: SeasonName
}

export interface NewTribe {
  tribeName: TribeName;
  tribeColor: TribeColor;
}

export type TribeEp = {
  tribeId: TribeId;
  tribeName: TribeName;
  tribeColor: TribeColor;
  episode: EpisodeNumber
};

export type TribeUpdate = {
  tribeColor: TribeColor;
  castaways: CastawayName[];
};